Refactor router modules
