+++ services/meta-data-repository/src/route/domains/schemas/get.js
@@ -1,0 +1,93 @@
+const express = require('express');
+const logger = require('@basaas/node-logger');
+const {
+    transformDbResults,
+    buildURI,
+} = require('../../../transform');
+const { SchemaDAO } = require('../../../dao');
+const { domainOwnerOrAllowed } = require('../../../middleware/permission');
+const Pagination = require('../../../util/pagination');
+const conf = require('../../../conf');
+const { isLocalRequest } = require('../../../util/common');
+
+const log = logger.getLogger(`${conf.logging.namespace}/domains/schemas/get`);
+
+const router = express.Router();
+
+router.get('/', domainOwnerOrAllowed({
+    permissions: ['not.defined'],
+}), async (req, res) => {
+    const pagination = new Pagination(
+        req.originalUrl,
+        SchemaDAO,
+    );
+
+    let data;
+    let meta;
+
+    if (req.hasAll) {
+        data = await SchemaDAO.findByDomain({
+            domainId: req.domainId,
+            options: pagination.props(),
+        });
+
+        meta = {
+            ...await pagination.calc({
+                domainId: req.domainId,
+            }),
+        };
+    } else {
+        data = await SchemaDAO.findByDomainAndEntity({
+            entityId: req.user.sub,
+            domainId: req.domainId,
+            options: pagination.props(),
+        });
+        meta = {
+            ...await pagination.calc({
+                domainId: req.domainId,
+                'owners.id': req.user.sub,
+            }),
+        };
+    }
+
+    res.send({
+        data: transformDbResults(data),
+        meta,
+    });
+});
+
+router.get('/:uri*', async (req, res, next) => {
+    if (!req.user && isLocalRequest(req)) {
+        return next();
+    }
+    domainOwnerOrAllowed({
+        permissions: ['not.defined'],
+    })(req, res, next);
+}, async (req, res, next) => {
+    try {
+        const schema = await SchemaDAO.findByURI({
+
+            uri: buildURI({
+                domainId: req.domainId,
+                uri: req.path.replace(/^\//, ''),
+            }),
+        });
+
+        if (!schema) return next({ status: 404 });
+
+        if (req.header('content-type') === 'application/schema+json') {
+            return res.send(schema.value);
+        }
+
+        res.send({
+            data: transformDbResults(schema),
+        });
+    } catch (err) {
+        log.error(err);
+        next({
+            status: 400,
+        });
+    }
+});
+
+module.exports = router;
